llm/requirements.txt
# Start of Selection
torch==2.7.1
transformers==4.52.4
pillow==11.2.1
datasets==3.6.0
accelerate==1.7.0
evaluate==0.4.3
bitsandbytes==0.46.0
trl==0.18.1
peft==0.15.2
matplotlib==3.10.3
tqdm==4.67.1
#flash-attn #cant have this for the finetuning for some reason
pydantic==2.11.5
openai==1.85.0
dotenv==0.9.9
tensorboard==2.19.0
torchvision==0.22.1
protobuf==6.31.1
sentencepiece==0.2.0
flask
pycocotools
scipy
scikit-image